+++ SocialSim/src/com/socialsim/model/simulator/mall/MallSimulator.java
@@ -43,9 +43,9 @@ public class MallSimulator extends Simulator {
     public static int totalAloneCount = 0;
     public static int totalCoupleCount = 0;
 
-    public static int averageNonverbalDuration = 0;
-    public static int averageCooperativeDuration = 0;
-    public static int averageExchangeDuration = 0;
+    public static float averageNonverbalDuration = 0;
+    public static float averageCooperativeDuration = 0;
+    public static float averageExchangeDuration = 0;
 
     public static int currentPatronPatronCount = 0;
     public static int currentPatronStaffStoreCount = 0;
